fix(postgres): cast EXTRACT(EPOCH) to float8 in prompt_received_seconds_ago
CI regression from PR #556 finding #1 (c215c4c2): Postgres's
EXTRACT(EPOCH FROM ...) always returns NUMERIC, never FLOAT8/double
precision, regardless of the operand types. sqlx's runtime `query()` API
tried to decode the "secs_ago" column straight into an `f64`, which
fails at runtime with a ColumnDecode mismatched-types error — not
caught by SQLX_OFFLINE since this uses the non-macro `sqlx::query()`
path (no compile-time query cache entry to go stale).

Fix: cast in the SQL itself (`::float8`) so Postgres returns the type
the Rust side already expects. No `.sqlx/` cache to regenerate — this
query doesn't use the `query!`/`query_as!` macros.

Verified against a live local Postgres:
`cargo nextest run -p engram-coordinator --test checkpoint_reconcile_live_pg
prompt_received_seconds_ago_resolves_by_prompt_id_and_misses_cleanly
--run-ignored all` now passes.

+++ crates/engram-postgres/src/lib.rs
@@ -2837,7 +2837,7 @@ impl MetadataStore for PostgresStore {
         // bias or drop samples.
         let row = sqlx::query(
             r#"
-            SELECT EXTRACT(EPOCH FROM (NOW() - created_at)) AS secs_ago
+            SELECT EXTRACT(EPOCH FROM (NOW() - created_at))::float8 AS secs_ago
               FROM session_events
              WHERE session_id = $1 AND kind = 'prompt_received' AND payload->>'prompt_id' = $2
              ORDER BY idx DESC
